stuff/vokabelnlernen_impl/parsing/requirements.txt
argparse==1.4.0
PyYAML==6.0
mysql-connector-python==8.0.30
paho-mqtt==1.6.1
webvtt-py==0.4.6